+++ modules/ml/src/test/java/org/apache/ignite/ml/selection/scoring/cursor/LocalLabelPairCursorTest.java
@@ -36,19 +36,19 @@ public class LocalLabelPairCursorTest {
         Map<Integer, double[]> data = new HashMap<>();
 
         for (int i = 0; i < 1000; i++)
-            data.put(i, new double[] { i, i});
+            data.put(i, new double[] {i, i});
 
         Vectorizer<Integer, double[], Integer, Double> vectorizer = new DoubleArrayVectorizer<Integer>().labeled(Vectorizer.LabelCoordinate.FIRST);
 
-        LabelPairCursor<Integer> cursor = new LocalLabelPairCursor<>(
+        LabelPairCursor<Double> cursor = new LocalLabelPairCursor<>(
             data,
             (k, v) -> v[1] % 2 == 0,
             vectorizer,
-            vec -> (int)vec.get(0)
+            vec -> vec.get(0)
         );
 
         int cnt = 0;
-        for (LabelPair<Integer> e : cursor) {
+        for (LabelPair<Double> e : cursor) {
             assertEquals(e.getPrediction(), e.getTruth());
             cnt++;
         }
